+++ server/frontend/src/stores/ui/index.js
@@ -4,6 +4,7 @@ import { useSplit } from './split'
 import { useTab } from './tab'
 import { useHelp } from './help'
 import { useNotification } from './notification'
+import { useMatchMode } from './matchMode'
 
 export const useUi = () => ({
   chart: useChart(),
@@ -11,5 +12,6 @@ export const useUi = () => ({
   split: useSplit(),
   tab: useTab(),
   help: useHelp(),
-  notification: useNotification()
+  notification: useNotification(),
+  matchMode: useMatchMode()
 })